requirements.txt
Flask==3.0.3
gunicorn==22.0.0
line-bot-sdk==3.14.2
pandas==2.2.2
numpy==2.0.1
pyarrow==16.1.0
matplotlib==3.9.0